refactor: organize trip generation code better
